Major refactoring of the Pattyn wrapper, mainly to off-load derivative computations to elsewhere in CISM.  The wrappers were growing to truly unmanageable proportions.
As a result of this, a lot more derivative fields were added to the glide_geomderv structure.
I am aware that this will increase memory usage somewhat, and am open to suggestions regarding
how this could be better handeled.  Hopefully other modules can make good use of the additional
fields.


git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism/glimmer/branches/lanl@1025 2cfc7cc4-626f-0410-8325-f6cb4dace9bf